% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTooltip.R
\name{addTooltip}
\alias{addTooltip}
\title{Add tooltip to an HTML element}
\usage{
addTooltip(el, tlp_color = "black", tlp_msg, tlp_position = "top")
}
\arguments{
\item{el}{Element for witch to add a tooltip.}

\item{tlp_color}{Tooltip Boostrap 5 color}

\item{tlp_msg}{Tooltip message}

\item{tlp_position}{Tooltip position}
}
\value{
A html element with tooltip
}
\description{
Add tooltip to an HTML element
}
