% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_gallery.R
\name{app_gallery}
\alias{app_gallery}
\title{Shiny app that demonstrates capabilities of ShinyRating package}
\usage{
app_gallery()
}
\value{
a shiny app demonstrating Interfacer capabilities
}
\description{
Shiny app that demonstrates capabilities of ShinyRating package
}
