% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_site.R
\name{app_site}
\alias{app_site}
\title{Shiny app that demonstrates capabilities of ShinyRating package}
\usage{
app_site()
}
\value{
a shiny app demonstrating Interfacer capabilities
}
\description{
Shiny app that demonstrates capabilities of ShinyRating package
}
