% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRating.R
\name{createRating}
\alias{createRating}
\title{Create rating icons}
\usage{
createRating(
  rating_id,
  i_name = "star",
  i_lib = "font-awesome",
  number = 10,
  init_value = 0,
  on_color = "primary",
  off_color = "dark",
  size = 2,
  anim = "none",
  duration = 2,
  cumul = FALSE,
  tlp = FALSE,
  tlp_msg = 1:number,
  tlp_position = "top",
  tlp_color = "black",
  read_only = FALSE,
  hover = FALSE
)
}
\arguments{
\item{rating_id}{The rating widget id.}

\item{i_name}{The icon names.}

\item{i_lib}{The icon library to use. Either "font-awesome" or "glyphicon".}

\item{number}{The number of icons in the rating widget.}

\item{init_value}{The Initial value of the rating widget.}

\item{on_color}{The icon colors when selected (hex or Bootstrap 5 color).}

\item{off_color}{The icon colors when deselected (hex or Bootstrap 5 color).}

\item{size}{The icon sizes (1 to 5).}

\item{anim}{The type of animations see \href{https://animate.style/}{Animate.css}.}

\item{duration}{The animation duration in seconds.}

\item{cumul}{Logical, whether or not to use cumulative color for the icons.}

\item{tlp}{Logical, whether or not to use a tooltip for the icon.}

\item{tlp_msg}{The character strings to be shown as message in tooltips.}

\item{tlp_position}{The tooltip positions, \code{"top"}, \code{"right"}, \code{"bottom"}, or \code{"left"}.}

\item{tlp_color}{The tooltip colors, \code{"primary"}, \code{"secondary"}, \code{"light"}, \code{"dark"}, \code{"info"}, \code{"warning"}, \code{"danger"}, \code{"success"}, \code{"white"}, or \code{"black"}.}

\item{read_only}{Logical, whether or not to use a read only mode for the icons.}

\item{hover}{Logical, whether or not to use a hover mode for the icons.}
}
\value{
A \code{shiny.tag} object for usage in \code{\link{ratingInput}}
}
\description{
Create rating icons for usage in \code{\link{ratingInput}}.
}
