% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Load package dependencies}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{libname}

\item{pkgname}{pkgname}
}
\value{
Load package dependencies
}
\description{
Load package dependencies
}
