% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_svg.R
\name{read_icon}
\alias{read_icon}
\title{load an svg file}
\usage{
read_icon(i_name, i_width = "2em", i_height = NULL)
}
\arguments{
\item{i_name}{svg path}

\item{i_width}{svg width}

\item{i_height}{svg height}
}
\value{
svg in tag
}
\description{
load an svg file
}
