% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gallery.R
\name{run_gallery}
\alias{run_gallery}
\title{Shiny app: shinyRating capabilities}
\usage{
run_gallery()
}
\value{
No returned value, this function just runs a Shiny app.
}
\description{
Shiny app demonstrating the rating widget capabilities.
}
