% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_character}
\alias{test_character}
\title{Test if values in a vector are a string}
\usage{
test_character(x)
}
\arguments{
\item{x}{Vector of values}
}
\value{
An error if at least one value is not a string
}
\description{
Test if values in a vector are a string
}
