% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_length}
\alias{test_length}
\title{Test if a vector has a length == 1 or equal to a given number}
\usage{
test_length(x, number)
}
\arguments{
\item{x}{Vector of values}
}
\value{
An error if the vector has a length != 1 and different to a given number
}
\description{
Test if a vector has a length == 1 or equal to a given number
}
