% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_logical}
\alias{test_logical}
\title{Test if values in a vector are logical}
\usage{
test_logical(x)
}
\arguments{
\item{x}{Vector of values}
}
\value{
An error if at least one value is not logical
}
\description{
Test if values in a vector are logical
}
