% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_na}
\alias{test_na}
\title{Test if values in a vector are NA}
\usage{
test_na(x)
}
\arguments{
\item{x}{Vector of values}
}
\value{
An error if at least one value is NA
}
\description{
Test if values in a vector are NA
}
