% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_null}
\alias{test_null}
\title{Test if a vector of values is null}
\usage{
test_null(x)
}
\arguments{
\item{x}{Vector of values}
}
\value{
An error if the vector is null
}
\description{
Test if a vector of values is null
}
