% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_pos_integer}
\alias{test_pos_integer}
\title{Test if values in a vector are positive integers}
\usage{
test_pos_integer(x, na_rm = FALSE)
}
\arguments{
\item{x}{Vector of values}

\item{na_rm}{Logical, are NA be removed ?}
}
\value{
An error if at least one value is not an integer
}
\description{
Test if values in a vector are positive integers
}
