% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIcon.R
\name{createIcon}
\alias{createIcon}
\title{Create an icon}
\usage{
createIcon(
  id,
  i_name = "star",
  i_lib = "font-awesome",
  i_number = 1,
  number = 10,
  rating_id,
  on_color = "primary",
  off_color = "dark",
  init_color = "dark",
  size = 2,
  anim = "none",
  duration = 2,
  cumul = FALSE,
  tlp = FALSE,
  tlp_msg = 1,
  tlp_position = "top",
  tlp_color = "black",
  read_only = FALSE,
  hover = FALSE
)
}
\arguments{
\item{id}{The icon id.}

\item{i_name}{The icon name.}

\item{i_lib}{The icon library to use. Either "font-awesome" or "glyphicon".}

\item{i_number}{The icon place in the rating widget.}

\item{number}{The total number of icons in the rating widget.}

\item{rating_id}{The rating widget id.}

\item{on_color}{The icon color when selected (hex or Bootstrap 5 color).}

\item{off_color}{The icon color when deselected (hex or Bootstrap 5 color).}

\item{init_color}{The icon initial color (hex or Bootstrap 5 color).}

\item{size}{The icon size (1 to 5).}

\item{anim}{The type of animation see \href{https://animate.style/}{Animate.css}.}

\item{duration}{The animation duration in seconds.}

\item{cumul}{Logical, whether or not to use cumulative color for the icon.}

\item{tlp}{Logical, whether or not to use a tooltip for the icon.}

\item{tlp_msg}{The character string to be shown as message in the tooltip.}

\item{tlp_position}{The tooltip position, \code{"top"}, \code{"right"}, \code{"bottom"}, or \code{"left"}.}

\item{tlp_color}{The tooltip color, \code{"primary"}, \code{"secondary"}, \code{"light"}, \code{"dark"}, \code{"info"}, \code{"warning"}, \code{"danger"}, \code{"success"}, \code{"white"}, or \code{"black"}.}

\item{read_only}{Logical, whether or not to use a read only mode for the icon.}

\item{hover}{Logical, whether or not to use a hover mode for the icon.}
}
\value{
A \code{shiny.tag} object for usage in \code{\link{createRating}}
}
\description{
Create an icon for usage in \code{\link{createRating}}.
}
