% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIcon.R
\name{createIcon}
\alias{createIcon}
\title{create an icon}
\usage{
createIcon(
  id,
  i_name = "star",
  i_lib = "font-awesome",
  i_number = 1,
  number = 10,
  rating_id,
  on_color = "primary",
  off_color = "dark",
  init_color = "dark",
  size = 2,
  anim = "none",
  duration = 2,
  cumul = TRUE,
  tlp = FALSE,
  tlp_msg = 1,
  tlp_position = "top",
  tlp_color = "black",
  read_only = FALSE
)
}
\arguments{
\item{id}{icon id}

\item{i_name}{Name of the icon.}

\item{i_lib}{The icon library to use. Either "font-awesome" or "glyphicon".}

\item{i_number}{Icon number}

\item{number}{Number of icons in rating input control}

\item{rating_id}{Rating input control id}

\item{on_color}{Selection color of icon (hex or Boostrap 5 color)}

\item{off_color}{Default color of icon (hex or Boostrap 5 color)}

\item{init_color}{Initial color of icon (hex or Boostrap 5 color)}

\item{size}{Size of the icons (1 to 5)}

\item{anim}{Type of animation from animate.css library (https://animate.style/)}

\item{duration}{Duration of the animation}

\item{cumul}{Cumulative color for icons (TRUE or FALSE)}

\item{tlp}{Tooltip for the icons (TRUE or FALSE)}

\item{tlp_msg}{Tooltip messages}

\item{tlp_position}{Tooltip positions}

\item{tlp_color}{Tooltip colors}

\item{read_only}{Read Only mode for icons (TRUE or FALSE)}
}
\value{
a rating icon
}
\description{
create an icon
}
\examples{
createIcon(id = "icon", rating_id = "rating_id")
}
