% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRating.R
\name{createRating}
\alias{createRating}
\title{Create rating icons}
\usage{
createRating(
  rating_id,
  i_name = "star",
  i_lib = "fontawesome",
  number = 10,
  init_value = 0,
  i_on_color = "primary",
  i_off_color = "dark",
  i_width = NULL,
  i_height = NULL,
  i_anim = "none",
  i_duration = 2,
  cumul = FALSE,
  tlp = FALSE,
  i_value = 1:number,
  i_margin_left = "auto",
  i_margin_right = "auto",
  tlp_position = "top",
  tlp_color = "black",
  read_only = FALSE,
  hover = FALSE
)
}
\arguments{
\item{rating_id}{The rating input id.}

\item{i_name}{The icon name (fontawesome) or path (local).}

\item{i_lib}{The icon library to use. Either "fontawesome" or "local".}

\item{number}{The number of icons in the rating input.}

\item{init_value}{The Initial value of the rating input.}

\item{i_on_color}{The icon colors when selected (hex or Bootstrap 5 color).}

\item{i_off_color}{The icon colors when deselected (hex or Bootstrap 5 color).}

\item{i_width}{The icons width in css units.}

\item{i_height}{The icons height in css units.}

\item{i_anim}{The type of animations see \href{https://animate.style/}{Animate.css}.}

\item{i_duration}{The animation duration in seconds.}

\item{cumul}{Logical, whether or not to use cumulative color for the icons.}

\item{tlp}{Logical, whether or not to use a tooltip for the icon.}

\item{i_value}{The character strings to be shown as message.}

\item{i_margin_left, i_margin_right}{The length value for the margin that's either
left or right of the icon. By default, \code{"auto"} is used for both
properties. If space is needed on either side then a length of \code{"0.2em"} is
recommended as a starting point.}

\item{tlp_position}{The tooltip positions, \code{"top"}, \code{"right"}, \code{"bottom"}, or \code{"left"}.}

\item{tlp_color}{The tooltip colors, \code{"primary"}, \code{"secondary"}, \code{"light"}, \code{"dark"}, \code{"info"}, \code{"warning"}, \code{"danger"}, \code{"success"}, \code{"white"}, or \code{"black"}.}

\item{read_only}{Logical, whether or not to use a read only mode for the icons.}

\item{hover}{Logical, whether or not to use a hover mode for the icons.}
}
\value{
A \code{shiny.tag} object for usage in \code{\link{ratingInput}}
}
\description{
Create rating icons for usage in \code{\link{ratingInput}}.
}
