% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRating.R
\name{createRating}
\alias{createRating}
\title{Create a icon rating}
\usage{
createRating(
  rating_id,
  i_name = "star",
  i_lib = "font-awesome",
  number = 10,
  init_value = 0,
  on_color = "primary",
  off_color = "dark",
  size = 2,
  anim = "none",
  duration = 2,
  cumul = TRUE,
  tlp = FALSE,
  tlp_msg = 1:number,
  tlp_position = "top",
  tlp_color = "black",
  read_only = FALSE,
  hover = FALSE
)
}
\arguments{
\item{rating_id}{Rating input control id}

\item{i_name}{Name of the icon.}

\item{i_lib}{The icon library to use. Either "font-awesome" or "glyphicon".}

\item{number}{Number of icons in rating input control.}

\item{init_value}{Initial value.}

\item{on_color}{Selection color of icons (hex or Boostrap 5 color)}

\item{off_color}{Default color of icons (hex or Boostrap 5 color)}

\item{size}{Size of the icons (1 to 5)}

\item{anim}{Type of animation from animate.css library (https://animate.style/)}

\item{duration}{Duration of the animation}

\item{cumul}{Cumulative color for icons (TRUE or FALSE)}

\item{tlp}{Tooltip for the icons (TRUE or FALSE)}

\item{tlp_msg}{Tooltip messages}

\item{tlp_position}{Tooltip positions}

\item{tlp_color}{Tooltip colors}

\item{read_only}{Read Only mode for icons (TRUE or FALSE)}

\item{hover}{Hover for icons (TRUE or FALSE)}
}
\value{
a icons rating
}
\description{
Create a icon rating
}
