% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratingInput.R
\name{ratingInput}
\alias{ratingInput}
\title{Create a rating input control}
\usage{
ratingInput(
  inputId,
  number = 5,
  on_color = "primary",
  off_color = "dark",
  label = NULL,
  size = 1,
  i_name = "cog",
  i_lib = "font-awesome",
  width = "auto",
  value = NA,
  anim = "none",
  duration = 2,
  cumul = TRUE,
  tlp = FALSE,
  tlp_msg = 1:number,
  tlp_position = "top",
  tlp_color = "black"
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{number}{Number of icons in rating input control}

\item{on_color}{Color of selected icons (hex or Boostrap 5 color)}

\item{off_color}{Color of unselected icons (hex or Boostrap 5 color)}

\item{label}{Display label for the control, or NULL for no label.}

\item{size}{Size of the icons (1 to 5)}

\item{i_name}{Names of the icons. Names with length 1 or equal to range from either Font Awesome (when i_lib="font-awesome") or Bootstrap Glyphicons (when i_lib="glyphicon") may be provided.}

\item{i_lib}{The icon library to use. Either "font-awesome" or "glyphicon".}

\item{width}{The width of the input, e.g. '400px', or '100\%'; see validateCssUnit()}

\item{value}{Initial value.}

\item{anim}{Type of animation from animate.css library (https://animate.style/)}

\item{duration}{Duration of the animation}

\item{cumul}{Cumulative color for icons (TRUE or FALSE)}

\item{tlp}{Tooltip for the icons (TRUE or FALSE)}

\item{tlp_msg}{Tooltip messages}

\item{tlp_position}{Tooltip positions}

\item{tlp_color}{Tooltip colors}
}
\value{
A numeric value of the rating input. The default value is NULL unless value is provided.
}
\description{
Create a rating input control
}
\examples{
ratingInput(inputId = "like", i_name = "heart", on_color = "danger")
}
