% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRatingInput.R
\name{updateRatingInput}
\alias{updateRatingInput}
\title{Change the value of a rating input on the client}
\usage{
updateRatingInput(session, inputId, value = NA)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer. Default is getDefaultReactiveDomain().}

\item{inputId}{The id of the input object.}

\item{value}{Initial value.}
}
\value{
an updated value of a rating input
}
\description{
Change the value of a rating input on the client
}
